# perf: Only check timeout every 256 files