<!--- Provide a general summary of your changes in the Title above -->
<!--- To help with semantic versioning the PR title should start with one of the conventional commit types. -->
<!--- The conventional commit types for Semantic PR are: feat, fix, docs, style, refactor, perf, test, build, ci, chore, revert -->

#### Description
<!--- Describe your changes in detail -->

When listing directory content, only check timeout every 256 files

#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
I noticed when listing directory content, the iterator will check system time against timeout for every single item. I thought changing it to check once in a while instead would be faster for large directories. I tried to compare timing locally for a large directory and it decreased the time by about 40-45% so I decided to open this PR to see if this is actually a good idea.

Note I actually use a stripped down fork from a quite old version (0.25.2) though it looks like this particular code hasn't changed that much between now and then. (but the log messages are different it seems)

I tested this change in a directory with 1685 files, without the change the logs read:

```
 TRACE starship::context             > Building a vector of directory files took 4.948ms
```

and with the change:

```
 TRACE starship::context             > Building a vector of directory files took 2.841ms

 TRACE starship::context             > Building a vector of directory files took 2.828ms

 TRACE starship::context             > Building a vector of directory files took 2.882ms

 TRACE starship::context             > Building a vector of directory files took 2.963ms
```

An alternative method is to use `.trailing_zeros()` instead of bitwise operations, which has the benefit of skipping the first entry as well, though the first interval is not consistent - say using `trailing_zeros() != 8` then the first check happens for index 256 (from 0) but after that it checks every 512 entries. Both methods provide similiar time though so I'm using the more "consistent" one in this PR.

I've also tried a bunch of other numbers than `8` and most others seem to provide (tiny bit) worse or rather identical result.

I'm not sure if there're existing benchmark for this particular scenario for `starship`? Or should I setup a more comprehensive test to show improvements?

#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, tests ran to see how -->
<!--- your change affects other areas of the code, etc. -->
- [x] I have tested using **MacOS**
- [ ] I have tested using **Linux**
- [ ] I have tested using **Windows**

#### Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] I have updated the documentation accordingly.
- [ ] I have updated the tests accordingly.