@davidkna 2020-07-21
The time-out was added for slow filesystems. I feel like 256 files is quite high for that. Maybe a smaller threshold is more appropriate?
You might also want to check if `% n` has the same speed as this bitwise operation, because it could be optimized for this special case. Also I'm not sure if this works on big endian (we had someone use starship on PowerPC).