@aquarhead 2020-07-21
> You might also want to check if % n has the same speed as this bitwise operation, because it could be optimized for this special case.

Wouldn't `%` be optimized into bitwise operations?

> Also I'm not sure if this works on big endian (we had someone use starship on PowerPC).

Switching to `.trailing_zeros()` might be better then? It seems [it uses CPU intrinsics](https://doc.rust-lang.org/src/core/num/mod.rs.html#2685-2687).

> I feel like 256 files is quite high for that.

I think doing `n.trailing_zeros() != 3` also works just fine with similiar timing, but I only have my local testing result, not a comprehensive benchmark to compare performance for all scenarios.