@davidkna 2020-07-21
> Wouldn't `%` be optimized into bitwise operations?

Maybe. It's more obvious what  `%` does and it should work on big endian while being just as efficient if it's optimized into bitwise operations.

> Switching to `.trailing_zeros()` might be better then? It seems [it uses CPU intrinsics](https://doc.rust-lang.org/src/core/num/mod.rs.html#2685-2687).

Not sure if it works. The docs talk about the number of trailing zeros in the "binary representation" which could refer to the in-memory representation.