@davidkna 2020-07-21
On second thought `& 0xFF` should work on both endians.