@whatisaphone 2020-07-28
That code should be using `Instant::now()`, not `SystemTime`. `SystemTime` is not monotonic, meaning the time could go backwards. The `unwrap()` in that code could panic if the system syncs with an NTP server, or daylight savings time happens, or the user is on a train traveling west, etc. [See docs](https://doc.rust-lang.org/std/time/struct.Instant.html)

And since this issue is about perf, I would guess that `Instant::now` is faster than `SystemTime::now`, since it doesn't have to sync with the system clock.