@andytom 2020-07-28
Thanks for the suggestion @whatisaphone I've created an issue so that this doesn't get forgotten about.